requirements.txt
asgiref==3.5.0
autopep8==1.6.0
Django==4.0.2
djangorestframework==3.13.1
pip==22.0.3
pycodestyle==2.8.0
pytz==2021.3
setuptools==60.9.0
sqlparse==0.4.2
toml==0.10.2
wheel==0.37.1
